+++ CHANGELOG.md
@@ -8,7 +8,12 @@ All notable changes to this project will be documented in this file.
 
 - Remove `resources` key from `DynamicValues` struct ([#734]).
 
+## Fixed
+
+- Fixed incorrect time calculation ([#735]).
+
 [#734]: https://github.com/stackabletech/operator-rs/pull/734
+[#735]: https://github.com/stackabletech/operator-rs/pull/735
 
 ## [0.64.0] - 2024-01-31